refactor: reorganize print template components
